+++ mxnet/src/main/java/org/bytedeco/javacpp/presets/mxnet.java
@@ -40,7 +40,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
     @Platform(value = {"linux-x86", "macosx"}, compiler = "cpp11", define = {"DMLC_USE_CXX11 1", "MSHADOW_USE_CBLAS 1", "MSHADOW_IN_CXX11 1"},
         include = {"mxnet/c_api.h", "mxnet/c_predict_api.h", /*"dmlc/base.h", "dmlc/io.h", "dmlc/logging.h", "dmlc/type_traits.h",
                    "dmlc/parameter.h", "mshadow/base.h", "mshadow/expression.h", "mshadow/tensor.h", "mxnet/base.h",*/},
-        link = "mxnet", includepath = {"/usr/local/cuda/include/",
+        link = "mxnet", resource = {"include", "lib"}, includepath = {"/usr/local/cuda/include/",
         "/System/Library/Frameworks/vecLib.framework/", "/System/Library/Frameworks/Accelerate.framework/"}, linkpath = "/usr/local/cuda/lib/") })
 public class mxnet implements InfoMapper {
     public void map(InfoMap infoMap) {